pyproject.toml
[project]
name = "umcp"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.115.12",
    "google-genai>=1.9.0",
    "langchain-community>=0.3.20",
    "langchain>=0.3.22",
    "pydantic>=2.11.1",
    "pydantic-settings>=2.8.1",
    "supabase>=2.15.0",
    "mcp[cli]>=1.6.0",
    "ipython>=9.0.2",
    "uvicorn>=0.34.2",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.6",
]
